[FIX] payment_alipay: fix the broken alipay tests
Backport of https://github.com/odoo/odoo/pull/34132

The test case of alipay payment is going to failed due to:

1) Now the redirect_url from the return_url is removed in commit :
https://github.com/odoo/odoo/commit/13512b0ca896ae7c3572145600fd6e846bbcac74
but it's still it's there in test cases so removed from test case.

2) To make the transaction done it must be in  draft, pending or authorised
state but in first test case payment process the transaction state is set to
cancel and in another payment process use the same transaction again to set
it success so before use the same transaction again just set the state to draft.

task- 2005926

closes odoo/odoo#34708

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>