[FIX] website_{profile, forum}: migrate events from forum to profile
After this commit  https://github.com/odoo/odoo/commit/6b73b7e4c575e44 some of the events are left
in website_forum, that caused some buttons to be inactive. In order to fix that
events moved to website_profile.

Note: code is moved as it is. Cleaning class / namespace will be done in master
to avoid issues due to stable version.

Related to task 2006045

closes odoo/odoo#34138

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>