[FIX] account: fix breadcrumbs, action panel which is not appearing in duplicating journal entry
Before this commit, when user duplicates the journal entry, the action panel and breadcrumbs are lost
due to 'inline' action given in it for opening form view in edit mode directly.

After this commit, when user duplicates the journal entry the action panel and breadcrumbs will be appear
and also the form view opens directly in the edit mode.

task-1999208

closes odoo/odoo#34127

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>