[FIX] web: resequence after draft (re-enable test)
When ebcb130729 was forward-ported, there was an adapatation in 12.3
version with 021995864 so the test pass (because of another issue where
if a line is moved after the draft line, the code tried to discard the
moved line and not the draft line).

The same adaptation is necessary in saas-12.4 but since the code changed
amply, the test was temporarily just skipped (in d6be3a34f9).

With this changeset, the adaptation is forward-ported from saas-12.3 to
saas-12.4 and the test is re-enabled.

opw-1966589
closes #34277

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>