[FIX] hw_escpos: Avoid crashes on error
Two different problems on the box and resulted in the crash of the Odoo
instance running on the IoT Box:

When `usblp` had troubles communicating with a printer, a NoDeviceError
was raised. The `printer` variable was still set from the last time the
loop was processed, so we tried to close it, resulting in an error.

The code handling generic execptions was broken.

closes odoo/odoo#34300

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>