[FIX] website_sale_delivery: adapt test to ea9138daea58f
There is a conflict in saas-12.3.
Indeed, the optional popup on a product when adding to cart got encapsulated
into a new module with ea9138daea58f.

But 97b3dc9344 introduced a new test in 12.0 about free delivery.
The tour expects the popup to be shown, but it is not the case anymore.

The tour needed to be adapted. Standard runbot was green since the test would
luckily be executed after `website_sale_product_configurator` would be
installed. That was not the case on runbot theme.

The chance is also taken to ensure the test is executed on the correct
delivery.

closes odoo/odoo#34367

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>