[FIX] base: restore address display alignments
Commit https://github.com/odoo/odoo/commit/c701fb62a02db524c0caaa71e516ca3ccdb6c303
used flex to properly align address information and the fa icons. The
problem is that using flex prevents the parent div to ask for the
address information to be center/right aligned. This is the case for
the job pages in internal.

This commit uses display: inline-block on the top-level element of the
address to solve the problem.

opw-2000947

closes odoo/odoo#34371

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>