[FIX] point_of_sale: Ask password after scaning badge
Since changes made on Promises in rev: da8ceb3fc270ca84840c32fa6a35dd01eff62eae

If a user scans his badge, it's pin code if he has any, is not asked,
and he's imediatly connected to it's employee.

To avoid this behavior, we are handeling correctly the Promise to avoid
this connection shortcut.

closes odoo/odoo#34449

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>